requirements/default.txt
Pillow==8.3.2
pycairo==1.20.1
pycountry==20.7.3
PyGObject==3.40.1
PyGObject-stubs==0.0.2
pytesseract==0.3.8
